Task 2 atomic

+++ src/SecondThread.java
@@ -14,8 +14,8 @@ public class SecondThread implements Runnable{
         Random random = new Random();
         int K2 = random.nextInt(10000, 20000);
         for (int i = 0; i < K2; i++) {
-            car.addToTotalSpent(random.nextDouble(0, 10));
-            user.addToBalance(random.nextDouble(0, 100));
+            car.addToTotalSpent(random.nextInt(0, 10));
+            user.addToBalance(random.nextInt(0, 100));
         }
     }
 }
